convert "nightlies" to "nightly"
